Move definition of vertex to entityhandle map
